requirements.txt
fastmcp==2.14.2
httpx
python-dotenv